agreement/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_migrate==4.0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.0
Flask-Cors==3.0.10
flasgger==0.9.7.1
